final changes

+++ src/app/components/manager-aside/manager-aside.component.ts
@@ -1,8 +1,7 @@
-import {Component, EventEmitter, Input, OnInit, Output, Renderer2} from '@angular/core';
-import {ManagerDataObject} from "../../entities/managerDataObject";
-import {EventsService} from "../../services/events.service";
-import {ManagerDataHandlerService} from "../../services/manager-data-handler.service";
-import {error} from "@angular/compiler/src/util";
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {ManagerDataObject} from '../../entities/managerDataObject';
+import {EventsService} from '../../services/events.service';
+import {ManagerDataHandlerService} from '../../services/manager-data-handler.service';
 
 @Component({
   selector: 'app-manager-aside',
@@ -12,32 +11,29 @@ import {error} from "@angular/compiler/src/util";
 export class ManagerAsideComponent implements OnInit {
   @Input() public tab: string = 'queries';
   @Input() public dataCollection: ManagerDataObject[] = [];
+  @Input() public isSavedToasterOn: boolean = false;
 
-  // @Output() public onRecordAdd: EventEmitter<ManagerDataObject> = new EventEmitter<ManagerDataObject>();
   @Output() public onSelect: EventEmitter<ManagerDataObject> = new EventEmitter<ManagerDataObject>();
 
-  // @Input() public dataCollection: ManagerDataObject[] = [new ManagerDataObject(1, 'Какой то вменяемый но оч длинный запрос', 'kek1', new Date(), 'New'), new ManagerDataObject(2, 'какой то там запрос', 'kek2', new Date(), 'Saved'), new ManagerDataObject(3, 'какой то там еще запрос', 'kek3', new Date(), 'Changed')];
-
-
-  // @ts-ignore
   public selectedItem: ManagerDataObject;
   public tempRecordId: number = 0;
 
-  constructor(private managerDataHandlerService: ManagerDataHandlerService, private eventsService: EventsService, private renderer: Renderer2) { }
+  constructor(private managerDataHandlerService: ManagerDataHandlerService, private eventsService: EventsService) { }
 
   ngOnInit(): void {
-
 
   }
 
   public deleteRecord(id: number): void {
-    this.managerDataHandlerService.deleteById(this.tab, id)
-      .subscribe((data) => {
-          console.log(data, ' Item deleted successfully');
-          this.eventsService.recordDeleted.emit(id);
-        },
-        error => console.log(error)
-      );
+    if (id < 0) {
+      this.eventsService.recordDeleted.emit(id);
+    } else if (id > 0) {
+      this.managerDataHandlerService.deleteById(this.tab, id)
+        .subscribe((data) => {
+            console.log(data, ' Item deleted successfully');
+            this.eventsService.recordDeleted.emit(id);
+          }, error => console.log(error));
+    }
   }
 
   public createNewTempRecord(): void {
@@ -52,7 +48,5 @@ export class ManagerAsideComponent implements OnInit {
     this.selectedItem = record;
     this.onSelect.emit(record);
   }
-
-
 
 }
